moved functions to header file
